+++ src/main/java/com/parzivail/pswm/mobs/MobDroidAstromechImperial2.java
@@ -64,7 +64,7 @@ public class MobDroidAstromechImperial2 extends EntityDroidBase
 	protected void entityInit()
 	{
 		super.entityInit();
-		dataWatcher.addObject(18, Byte.valueOf((byte)0));
+		dataWatcher.addObject(18, (byte)0);
 	}
 
 	@Override
