+++ src/main/java/com/parzivail/pswm/mobs/MobDroidProbe.java
@@ -83,7 +83,7 @@ public class MobDroidProbe extends EntityDroidBase implements IShootThings
 	protected void entityInit()
 	{
 		super.entityInit();
-		dataWatcher.addObject(18, Byte.valueOf((byte)0));
+		dataWatcher.addObject(18, (byte)0);
 	}
 
 	@Override
