+++ src/main/java/com/parzivail/pswm/mobs/MobDroidSurgical.java
@@ -65,7 +65,7 @@ public class MobDroidSurgical extends EntityDroidBase
 	protected void entityInit()
 	{
 		super.entityInit();
-		dataWatcher.addObject(18, Byte.valueOf((byte)0));
+		dataWatcher.addObject(18, (byte)0);
 	}
 
 	@Override
