+++ src/main/java/com/parzivail/pswm/vehicles/VehicSkyhopper.java
@@ -28,7 +28,7 @@ public class VehicSkyhopper extends VehicleAirBase
 	public void entityInit()
 	{
 		super.entityInit();
-		this.dataWatcher.addObject(WING_DW, Float.valueOf(0.8f));
+		this.dataWatcher.addObject(WING_DW, 0.8f);
 		this.dataWatcher.setObjectWatched(WING_DW);
 	}
