+++ src/main/java/com/parzivail/util/block/ItemMultiHeightBlock.java
@@ -41,7 +41,7 @@ public class ItemMultiHeightBlock extends ItemBlockWithMetadata
 
 				if (j1 < 15 && world.checkNoEntityCollision(this.field_150939_a.getCollisionBoundingBoxFromPool(world, x, y, z)) && world.setBlockMetadataWithNotify(x, y, z, j1 + 1 | i1 & -8, 2))
 				{
-					world.playSoundEffect((double)((float)x + 0.5F), (double)((float)y + 0.5F), (double)((float)z + 0.5F), this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
+					world.playSoundEffect((float)x + 0.5F, (float)y + 0.5F, (float)z + 0.5F, this.field_150939_a.stepSound.func_150496_b(), (this.field_150939_a.stepSound.getVolume() + 1.0F) / 2.0F, this.field_150939_a.stepSound.getPitch() * 0.8F);
 					--stack.stackSize;
 					return true;
 				}
